using size_t

+++ ForQuizz/Source.cpp
@@ -1,4 +1,5 @@
 #include <iostream>
+#include <cstddef>
 
 using namespace std;
 
@@ -7,7 +8,7 @@ int main()
     float input;
     float sum = 0;
 
-    for (int i = 0; i < 5; i++)
+    for (size_t i = 0; i < 5; i++)
     {
 	   cout << "What is the next number? " << endl;
 	   cin >> input;
